fixed bug where ticket could be closed twice

+++ src/support/feature.js
@@ -351,13 +351,18 @@ class SupportFeature {
 
     async closeTicket(channel, ticket, executor) {
 
+        const statusName = ticket?.status?.name
         const closer = (executor?.id) ? { closer: { discord_id: executor.id } } : { id_closer: null }
         await this.database.tickets.update({ id_ticket: ticket.id_ticket }, { status: { name: "deleted" }, ...closer })
 
-        this.database.ipc.notify('ticket_closed', { guild_id: channel.guild.id, ticket, executor_id: (executor?.id ?? null) })
-        await this.transcriber.send(channel.guild, ticket)
-  
-        if (typeof channel.delete === "function") await channel.delete().catch(() => { /* Channel could already be deleted. */ })
+        if (!statusName || statusName !== 'deleted') {
+
+            this.database.ipc.notify('ticket_closed', { guild_id: channel.guild.id, ticket, executor_id: (executor?.id ?? null) })
+            await this.transcriber.send(channel.guild, ticket)
+      
+            if (typeof channel.delete === "function") await channel.delete().catch(() => { /* Channel could already be deleted. */ })
+
+        }
 
     }
